Add index.ts, rename gsSpec.ts to toplevel.ts. (shouldn't edit `ggschema` when refactoring ggvega TS)
