+++ inst/js/ggvega.js
@@ -1618,24 +1618,16 @@
     		type: "object"
     	},
     	ScaleType: {
-    		description: "The type of scale.  Vega-Lite supports the following categories of scale types:\r\n\r\n1) [**Continuous Scales**](https://vega.github.io/vega-lite/docs/scale.html#continuous)\r\n-- mapping continuous domains to continuous output ranges\r\n([`\"linear\"`](https://vega.github.io/vega-lite/docs/scale.html#linear),\r\n[`\"pow\"`](https://vega.github.io/vega-lite/docs/scale.html#pow),\r\n[`\"sqrt\"`](https://vega.github.io/vega-lite/docs/scale.html#sqrt),\r\n[`\"symlog\"`](https://vega.github.io/vega-lite/docs/scale.html#symlog),\r\n[`\"log\"`](https://vega.github.io/vega-lite/docs/scale.html#log),\r\n[`\"time\"`](https://vega.github.io/vega-lite/docs/scale.html#time),\r\n[`\"utc\"`](https://vega.github.io/vega-lite/docs/scale.html#utc).\r\n\r\n2) [**Discrete Scales**](https://vega.github.io/vega-lite/docs/scale.html#discrete) --\r\nmapping discrete domains to discrete\r\n([`\"ordinal\"`](https://vega.github.io/vega-lite/docs/scale.html#ordinal)) or continuous\r\n([`\"band\"`](https://vega.github.io/vega-lite/docs/scale.html#band) and\r\n[`\"point\"`](https://vega.github.io/vega-lite/docs/scale.html#point)) output ranges.\r\n\r\n3) [**Discretizing\r\nScales**](https://vega.github.io/vega-lite/docs/scale.html#discretizing) -- mapping\r\ncontinuous domains to discrete output ranges\r\n[`\"bin-ordinal\"`](https://vega.github.io/vega-lite/docs/scale.html#bin-ordinal),\r\n[`\"quantile\"`](https://vega.github.io/vega-lite/docs/scale.html#quantile),\r\n[`\"quantize\"`](https://vega.github.io/vega-lite/docs/scale.html#quantize) and\r\n[`\"threshold\"`](https://vega.github.io/vega-lite/docs/scale.html#threshold).\r\n\r\n__Default value:__ please see the [scale type\r\ntable](https://vega.github.io/vega-lite/docs/scale.html#type).",
     		"enum": [
-    			"band",
-    			"bin-ordinal",
-    			"linear",
-    			"log",
-    			"ordinal",
-    			"point",
-    			"pow",
-    			"quantile",
-    			"quantize",
-    			"sqrt",
-    			"symlog",
-    			"threshold",
-    			"time",
-    			"utc"
+    			"log"
     		],
     		type: "string"
+    	},
+    	Scales: {
+    		items: {
+    			$ref: "#/definitions/Scale"
+    		},
+    		type: "array"
     	},
     	StandardType: {
     		"enum": [
@@ -1671,10 +1663,7 @@
     				$ref: "#/definitions/Layers"
     			},
     			scales: {
-    				items: {
-    					$ref: "#/definitions/Scale"
-    				},
-    				type: "array"
+    				$ref: "#/definitions/Scales"
     			}
     		},
     		required: [
