+++ src-ext/Typescript/ggschema/src/index.ts
@@ -1,0 +1,6 @@
+export {TopLevelSpec} from './toplevel';
+
+export * from './data';
+export * from './layer';
+export * from './labels';
+export * from './scale';
