+++ src-ext/Typescript/ggschema/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "files": ["src/gsSpec.ts"],
+  "files": ["src/index.ts"],
   "compilerOptions": {
     /* Basic Options */
     // "incremental": true,                   /* Enable incremental compilation */
